requirements.txt
fpdf==1.7.2
Pillow==9.2.0
PyPDF2==2.11.1
typing_extensions==4.4.0
